+++ app/components/impulse/select_component.rb
@@ -23,15 +23,31 @@ module Impulse
 
     def options_from_choices
       return [] if @choices.nil?
-      @choices.map do |choice|
+      return grouped_options_for_select(@choices) if grouped_choices?
+      options_for_select(@choices)
+    end
+
+    def options_for_select(choices)
+      choices.map do |choice|
         html_attributes = option_html_attributes(choice)
         text, value = option_text_and_value(choice)
         OptionStruct.new(value, text, html_attributes)
       end
     end
 
+    def grouped_options_for_select(grouped_choices)
+      grouped_choices.map do |grouped_choice|
+        title, container = grouped_choice
+        [title, options_for_select(container)]
+      end
+    end
+
     def find_option(value)
-      (options.presence || options_from_choices).find { |option| option.value.to_s == value.to_s }
+      (options.presence || options_from_choices).flatten.find { |option| option.try(:value)&.to_s == value.to_s }
+    end
+
+    def grouped_choices?
+      @choices.present? && @choices.first.respond_to?(:last) && Array === @choices.first.last
     end
 
     def multiple?